clean up rpc handlers a bit
